adds a helpful comment and two missing semicolons
The new comment links to the google api console. It took me longer to
find that than it did to get these examples up and running.

The semicolons weren't critical, but it's too easy to miss their lack
if someone is copy/pasting chunks of this code into their own project.