OAuth 2.0 Scope
OAuth 2.0 requires a scope parameter to work, just like OAuth 1.  I used a different URL in OAuth 2, as I seemed to have problems using https://www.google.com/m8/feeds.  It might be a local problem, but the url I suggested might be better suited for basic use-cases.

I also provided links to the Google documentation on scope.  It's not great, but it helps.